More Minor Changes
The app now under GPL-3.0 !!
