+++ src/pages/adhkar.html
@@ -83,7 +83,8 @@
 
         </div>
 
-        <div id="adhkar" onmousedown='return false;' onselectstart='return false;' class="animate__animated animate__fadeIn">
+        <div id="adhkar" onmousedown='return false;' onselectstart='return false;'
+            class="animate__animated animate__fadeIn">
 
             <div id="title">
