+++ src/pages/adhkar/evening.html
@@ -51,7 +51,7 @@
             <a href="../home.html">
                 <img src="../../public/icon/home.png" class="icon_menu" id="icon_home">
             </a>
-            
+
             <a href="../surah.html">
                 <img src="../../public/icon/quran.png" class="icon_menu" id="icon_menu_surah">
             </a>
@@ -625,7 +625,7 @@
         </div>
 
     </div>
-
+    <script src="../public/js/adhkar.js"></script>
 </body>
 
 </html>
