+++ src/pages/adhkar/food.html
@@ -156,7 +156,9 @@
 
             <div class="adhkar" id="5">
                 <p class="text">
-                    كَانَ صَلَّى اللهُ عَلَيْهِ وَسَلَّمَ يَشْرَبُ فِي ثَلاَثَةِ أَنْفَاسٍ، إِذَا أَدْنَى الإِنَاءَ إِلَى فَمِهِ سَمَّى اللهَ تَعَالَى, وَإِذَا أَخَّرَهُ حَمِدَ اللهَ تَعَالَى، يَفْعَلُ ذَلِكَ ثَلاَثَ مَرَّاتٍ 
+                    كَانَ صَلَّى اللهُ عَلَيْهِ وَسَلَّمَ يَشْرَبُ فِي ثَلاَثَةِ أَنْفَاسٍ، إِذَا أَدْنَى الإِنَاءَ
+                    إِلَى فَمِهِ سَمَّى اللهَ تَعَالَى, وَإِذَا أَخَّرَهُ حَمِدَ اللهَ تَعَالَى، يَفْعَلُ ذَلِكَ ثَلاَثَ
+                    مَرَّاتٍ
                 </p>
                 <p class="text_bottom">هدى النبى فى الشرب</p>
                 <ul class="copy_and_paste">
@@ -173,7 +175,7 @@
         </div>
 
     </div>
-
+    <script src="../public/js/adhkar.js"></script>
 </body>
 
 </html>
