+++ src/pages/adhkar/prayer.html
@@ -51,7 +51,7 @@
             <a href="../home.html">
                 <img src="../../public/icon/home.png" class="icon_menu" id="icon_home">
             </a>
-            
+
             <a href="../surah.html">
                 <img src="../../public/icon/quran.png" class="icon_menu" id="icon_menu_top">
             </a>
@@ -85,7 +85,7 @@
 
         <div id="prayer" class="animate__animated animate__fadeIn">
 
-            <div class="adhkar" id="1" >
+            <div class="adhkar" id="1">
                 <p class="text">
                     اللهم بَاعِدْ بَيْنِيْ وَبَيْنَ خَطَايَايَ كَمَا بَاعَدْتَ بَيْنَ الْمَشْرِقِ وَالْمَغْرِبِ، اللهم
                     نَقِّنِيْ مِنْ خَطَايَايَ كَمَا يُنَقَّى الثَّوْبُ الْأبْيَضُ مِنَ الدَّنَسِ، اللهم اغْسِلْنِيْ مِنْ
@@ -345,6 +345,7 @@
         </div>
 
     </div>
+    <script src="../public/js/adhkar.js"></script>
 
 </body>
