+++ src/pages/hisnmuslim.html
@@ -44,7 +44,7 @@
             <a href="./home.html">
                 <img src="../public/icon/home.png" class="icon_menu" id="icon_home">
             </a>
-            
+
             <a href="./surah.html">
                 <img src="../public/icon/quran.png" class="icon_menu" id="icon_surah">
             </a>
@@ -86,7 +86,8 @@
 
         </div>
 
-        <div id="hisnmuslim" onmousedown='return false;' onselectstart='return false;' class="animate__animated animate__fadeIn">
+        <div id="hisnmuslim" onmousedown='return false;' onselectstart='return false;'
+            class="animate__animated animate__fadeIn">
 
             <div id="hisn">
                 <h1 id="titel_hisn">حصن المسلم</h1>
