+++ src/pages/home.html
@@ -81,7 +81,7 @@
             <img src="../public/icon/logo.png" id="logo">
 
             <ul id="pages">
-                
+
                 <li id="page_quran">
                     <p class="page">القرآن الكريم</p>
                     <!-- <img src="../public/icon/quran.png" id="page_icon"> -->
@@ -115,7 +115,7 @@
             </ul>
         </div>
     </div>
-
+    <script src="../public/js/home.js"></script>
 </body>
 
 </html>
