+++ src/pages/quran.html
@@ -51,7 +51,7 @@
             <a href="./home.html">
                 <img src="../public/icon/home.png" class="icon_menu" id="icon_home">
             </a>
-            
+
             <a href="../pages/surah.html">
                 <img src="../public/icon/quran.png" class="icon_menu" id="icon_surah">
             </a>
