+++ src/pages/quran_mp3.html
@@ -44,7 +44,7 @@
             <a href="./home.html">
                 <img src="../public/icon/home.png" class="icon_menu" id="icon_home">
             </a>
-            
+
             <a href="./surah.html">
                 <img src="../public/icon/quran.png" class="icon_menu" id="icon_surah">
             </a>
@@ -76,7 +76,8 @@
 
         </div>
 
-        <div id="quran_mp3" onmousedown='return false;' onselectstart='return false;' class="animate__animated animate__fadeIn">
+        <div id="quran_mp3" onmousedown='return false;' onselectstart='return false;'
+            class="animate__animated animate__fadeIn">
 
             <ul id="Sheikh"></ul>
             <ul id="ul_group"></ul>
@@ -84,7 +85,6 @@
         </div>
 
     </div>
-
 </body>
 
 </html>
