+++ src/pages/surah.html
@@ -44,11 +44,11 @@
             <a href="./home.html">
                 <img src="../public/icon/home.png" class="icon_menu" id="icon_home">
             </a>
-            
+
             <a href="./surah.html">
                 <img src="../public/icon/quran.png" class="icon_menu" id="icon_surah">
             </a>
-            
+
             <a href="./quran_mp3.html">
                 <img src="../public/icon/sound.png" class="icon_menu" id="icon_menu_mp3">
             </a>
