+++ src/preload/preload_modules/quran.js
@@ -1,4 +1,4 @@
-module.exports = function Quran(fs, path, App_Path, copy) {
+module.exports = function Quran(fs, path, App_Path) {
 
     let Quran_json = fs.readJsonSync(path.join(__dirname, '../../data/Quran.json'));
     let Now = fs.readJsonSync(path.join(App_Path, './data/Now.json'));
@@ -10,7 +10,14 @@ module.exports = function Quran(fs, path, App_Path, copy) {
     document.getElementById('Surah_text').innerText = ArrayQuran?.Surah;
     document.getElementById('Surah_text').addEventListener('click', (e) => {
         document.getElementById('alrt').style = 'display:inline-flex;'
-        copy('Surah_text');
+        
+        let createRange = document.createRange();
+        createRange.selectNode(document.getElementById('Surah_text'));
+        window.getSelection().removeAllRanges();
+        window.getSelection().addRange(createRange);
+        document.execCommand('copy');
+        window.getSelection().removeAllRanges();
+
         setTimeout(() => {
 
             document.getElementById('Surah_text').style = 'cursor: auto;'
